Update HelloToast example project to match updated lesson 1.1
