--- HelloToast/app/src/main/res/values/ic_launcher_background.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="ic_launcher_background">#D74B43</color>
-</resources>
