+++ HelloToast/app/src/main/res/values/strings.xml
@@ -1,23 +1,7 @@
-<!--
-  ~ Copyright (C) 2018 Google Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
 <resources>
-    <string name="app_name">Hello Toast</string>
-    <string name="button_label_toast">Toast</string>
-    <string name="button_label_count">Count</string>
-    <string name="count_initial_value">0</string>
-    <string name="toast_message">Hello Toast!</string>
+  <string name="app_name">Hello Toast</string>
+  <string name="button_label_toast">Toast</string>
+  <string name="button_label_count">Count</string>
+  <string name="count_initial_value">0</string>
+  <string name="toast_message">Hello Toast!</string>
 </resources>
